Acquire the default PhysX Fixed Time Step and apply the m_timeDistortionFactor to it every time
This ensures that both the simulation time scale and the PhysX Fixed
Time Step are adjusted by the same factor each time. Previously, each

This ensures that both the simulation time scale and the PhysX Fixed
Time Step are adjusted by the same factor each time. Previously, each
time the PhysX Fixed Time Step was changed it retained its previous
value and was multiplied against the new factor that was given.